% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{download_player_game_plot}
\alias{download_player_game_plot}
\title{Download player metric plot}
\usage{
download_player_game_plot(
  input_df,
  input_next_game,
  input_season,
  input_weeks,
  input_player,
  input_metric,
  input_threshold
)
}
\arguments{
\item{input_df}{data frame}

\item{input_next_game}{data frame with future games}

\item{input_season}{players to filter}

\item{input_weeks}{positions to filter}

\item{input_player}{metrics to filter}

\item{input_metric}{minimum weeks played}

\item{input_threshold}{minimum weeks played}
}
\value{
Download player metric plot
}
\description{
Download player metric plot
}
