% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{output_player_game_plot}
\alias{output_player_game_plot}
\title{Player metric plot}
\usage{
output_player_game_plot(
  input_df,
  input_season,
  input_weeks,
  input_player,
  input_metric,
  input_threshold
)
}
\arguments{
\item{input_df}{data frame}

\item{input_season}{players to filter}

\item{input_weeks}{positions to filter}

\item{input_player}{metrics to filter}

\item{input_metric}{minimum weeks played}

\item{input_threshold}{minimum weeks played}
}
\value{
Player metric plot
}
\description{
Player metric plot
}
